move base classes to separate files (avoid circular import)
