+++ anomalib/data/mvtec.py
@@ -32,8 +32,8 @@ from urllib.request import urlretrieve
 import albumentations as A
 from pandas import DataFrame
 
-from anomalib.data.base import AnomalibDataModule, AnomalibDataset, Split, ValSplitMode
-from anomalib.data.utils import DownloadProgressBar, hash_check
+from anomalib.data import AnomalibDataModule, AnomalibDataset
+from anomalib.data.utils import DownloadProgressBar, Split, ValSplitMode, hash_check
 from anomalib.pre_processing import PreProcessor
 
 logger = logging.getLogger(__name__)
