+++ anomalib/data/utils/__init__.py
@@ -11,7 +11,7 @@ from .image import (
     get_image_height_and_width,
     read_image,
 )
-from .split import concatenate_datasets, random_split
+from .split import Split, ValSplitMode, concatenate_datasets, random_split
 
 __all__ = [
     "generate_output_image_filename",
@@ -23,4 +23,6 @@ __all__ = [
     "DownloadProgressBar",
     "random_split",
     "concatenate_datasets",
+    "Split",
+    "ValSplitMode",
 ]
